handle executors appearing as child or parent proc
Some injects may be only detected but not prevented, resulting in an
"evidence" that may show the signature process name as a child
"process_name" or parent process.

We cannot yet build a logical collection of AND/OR required or optional
signatures on expectations, so executors may define a "process_name"
meaning any process being a child or a parent.

Signed-off-by: Antoine MAZEAS <antoine.mazeas@filigran.io>