align requirements.txt to branch practice
note this will not provide the correct pyobas version for a lot of cases
in development work whenever there is a dependency on an unreleased
version of pyobas.

To work around this, individual developers should manually checkout and install the
unreleased pyobas in their environment after installing all other requirements.

Signed-off-by: Antoine MAZEAS <antoine.mazeas@filigran.io>